Implemented drag-and-drop functionality with animations and reset feature.
Drag-and-Drop Targets:

Added multiple target boxes that accept drag-and-drop events.
Implemented functionality to toggle animations upon successful drop.
Animated Visibility:

Introduced scaling and fading animations for active target boxes.
Object Movement:

Implemented a state-driven mechanism to update the position and direction of the draggable object.
Rotation Animation:

Applied a repeating rotation animation when in the "playing" state.
Reset Functionality:

Added a button to reset the position of the draggable object to the center.
Improved UI